@@ -253,10 +253,10 @@ public class ConfigurationOverrider {
                 overrideReport.put("force scan", sf.toString());
             });
 
-            Optional.ofNullable(s.isPublicScan()).ifPresent(sf -> {
-                request.setPublicScan(sf);
-                overrideReport.put("public scan", sf.toString());
-            });
+//            Optional.ofNullable(s.isPublicScan()).ifPresent(sf -> {
+//                request.setPublicScan(sf);
+//                overrideReport.put("public scan", sf.toString());
+//            });
 
             Optional.ofNullable(s.getPreset()).ifPresent(sp -> {
                 request.setScanPreset(sp);
